move Client into client/ directory
