+++ client/route_in.go
@@ -1,0 +1,8 @@
+package client
+
+// ------ Incoming Requests ---------- //
+
+// InMsgHandler registers a handler to accept incoming messages from the server.
+func (c *Client) InMsgHandler(handler func(m []Message) error) {
+	c.inMsgHandler = handler
+}
